Remove hermitian from the assumptions system
This commit definitely removes hermitian and antihermitian from the
assumptions system. For example, there are no more facts that use
Q.hermitian or you can't define a symbol that is 'hermitian' with the
argument 'hermitian=True'.
It also modifies or removes some tests that use the hermitian
assumptions.

This commit